Split off all of the map-handling functions and variables of the main view controller into a superclass, ParkMapController.
